Adding Help link to Dashboard in order to display instructions
