+++ src/shared/SimpleHeaderSection.js
@@ -4,11 +4,33 @@ import { Link } from 'react-router-dom';
 import './SimpleHeaderSection.css';
 import Logo from './Logo.js';
 import iconUserDefault from '../images/general/icon-user-default.svg';
+import InstructionsModal from './InstructionsModal';
 
 class SimpleHeaderSection extends Component {
+  constructor() {
+    super();
+
+    this.state = {
+      showInstructionsModal: false
+    };
+
+    this.toggleInstructionsModal = this.toggleInstructionsModal.bind(this);
+  }
+
+  toggleInstructionsModal() {
+    const currentValue = this.state.showInstructionsModal;
+    this.setState({ showInstructionsModal: !currentValue });
+  }
+
   renderInfoSection() {
     return (
       <>
+        {this.state.showInstructionsModal ? (
+          <InstructionsModal
+            visible={this.state.showInstructionsModal}
+            toggle={this.toggleInstructionsModal}
+          />
+        ) : ''}
         <div className="userSection">
           <img
             className="userIcon noselect"
@@ -21,6 +43,9 @@ class SimpleHeaderSection extends Component {
           <Link to="/leaderboard" className="leaderboardButton">
             Leaderboard
           </Link>
+          <button className="link helpButton" onClick={this.toggleInstructionsModal}>
+            Help
+          </button>
           <a
             href="https://www.crowds-cure.org/privacy"
             target="privacy"
@@ -28,9 +53,9 @@ class SimpleHeaderSection extends Component {
           >
             Privacy
           </a>
-          <span className="logoutButton" onClick={this.logout}>
+          <button className="link logoutButton" onClick={this.logout}>
             Logout
-          </span>
+          </button>
         </div>
       </>
     );
